Fix circular import by moving wired_limit to stt/utils.py
