+++ mlx_audio/stt/models/voxtral/voxtral.py
@@ -10,8 +10,7 @@ import mlx.nn as nn
 import numpy as np
 from tqdm import tqdm
 
-from mlx_audio.stt.generate import wired_limit
-from mlx_audio.stt.utils import get_model_path
+from mlx_audio.stt.utils import get_model_path, wired_limit
 
 from ..base import STTOutput
 from .config import AudioConfig, ModelConfig
